Ensure shift state maintained if a numeric character is typed
The HID "write" function sends shift up and down codes when sending some characters such as '*'. After writing any numeric we now check to see if Shift is locked, and if it is re-send a shift key down code.

Also removed some debug in the Extra shift character send code.